collision fix

+++ src/simulation.h
@@ -27,7 +27,7 @@
 enum REB_BINARYDATA_ERROR_CODE; // Forward declaration.
 
 // Finds the two largest particles in the simulation. *p1 and *p2 will be set to the indices of the largest particles.
-void reb_simulation_two_largest_particles(struct reb_simulation* r, int* p1, int* p2);
+void reb_simulation_two_largest_particles(struct reb_simulation* r, size_t* p1, size_t* p2);
 
 // Only free memory in pointers of a simulation, but not the simulation itself.
 DLLEXPORT void reb_simulation_free_pointers(struct reb_simulation* const r);
